requirements.txt
numpy==1.23.4
pandas==1.5.1
scikit-learn==1.1.3
pytest==7.2.0
jupyter==1.0.0
jupyterlab==3.5.0
fastapi==0.86.0
uvicorn==0.19.0
dvc-s3==2.21.0
